bruh

+++ include/sdl.hpp
@@ -2,5 +2,5 @@ struct SDL{
     static void Present();
     static void Present(const char* message);
     static void Clear();
-    static void Dot(int x, int y, const char* mark);
+    static void Dot(float x, float y, const char* mark);
 };
